refactor: move recent messages to a separate component
